requirements.txt
python-telegram-bot==22.1
python-telegram-bot[job-queue]
requests==2.31.0
qbittorrent-api==v2025.5.0
